+++ database/migrations/2019_01_01_000000_add_disenio_to_en_disenio.php
@@ -14,7 +14,7 @@ class AddDisenioToEnDisenio extends Migration
     public function up()
     {
         Schema::table('en_disenios', function (Blueprint $table) {
-            $table->integer('disenios_id')->unsigned();
+            $table->unsignedBigInteger('disenios_id')->unsigned();
             $table->foreign('disenios_id')->references('id')->on('disenios');
         });
     }
