+++ database/migrations/2019_01_01_000000_add_jugadores_to_disenios_jugadores.php
@@ -14,7 +14,7 @@ class AddJugadoresToDiseniosJugadores extends Migration
     public function up()
     {
         Schema::table('disenios_jugadores', function (Blueprint $table) {
-            $table->integer('jugadores_id')->unsigned();
+            $table->unsignedBigInteger('jugadores_id')->unsigned();
             $table->foreign('jugadores_id')->references('id')->on('jugadores');
         });
     }
