+++ database/migrations/2019_01_01_000000_add_planetas_to_en_disenio.php
@@ -14,7 +14,7 @@ class AddPlanetasToEnDisenio extends Migration
     public function up()
     {
         Schema::table('en_disenios', function (Blueprint $table) {
-            $table->integer('planetas_id')->unsigned();
+            $table->unsignedBigInteger('planetas_id')->unsigned();
             $table->foreign('planetas_id')->references('id')->on('planetas');
         });
     }
