+++ database/migrations/2019_01_01_000000_add_planetas_to_en_disenio_en_planetas.php
@@ -14,7 +14,7 @@ class AddPlanetasToEnDisenioEnPlanetas extends Migration
     public function up()
     {
         Schema::table('disenios_en_planetas', function (Blueprint $table) {
-            $table->integer('planetas_id')->unsigned();
+            $table->unsignedBigInteger('planetas_id')->unsigned();
             $table->foreign('planetas_id')->references('id')->on('planetas');
         });
     }
