+++ database/migrations/2019_03_01_000000_add_construciones_id_to_planetas.php
@@ -14,7 +14,7 @@ class AddConstrucionesIdToPlanetas extends Migration
     public function up()
     {
         Schema::table('construcciones', function (Blueprint $table) {
-            $table->integer('planetas_id')->unsigned();
+            $table->unsignedBigInteger('planetas_id')->unsigned();
             $table->foreign('planetas_id')->references('id')->on('planetas');
         });
     }
