+++ database/migrations/2019_03_01_000000_add_jugadores_id_to_investigaciones.php
@@ -14,7 +14,7 @@ class AddJugadoresIdToInvestigaciones extends Migration
     public function up()
     {
         Schema::table('investigaciones', function (Blueprint $table) {
-            $table->integer('jugadores_id')->unsigned();
+            $table->unsignedBigInteger('jugadores_id')->unsigned();
             $table->foreign('jugadores_id')->references('id')->on('jugadores');
         });
     }
