+++ database/migrations/2019_07_01_000000_add_industrias_id_to_planetas.php
@@ -14,7 +14,7 @@ class AddIndustriasIdToPlanetas extends Migration
     public function up()
     {
         Schema::table('industrias', function (Blueprint $table) {
-            $table->integer('planetas_id')->unsigned();
+            $table->unsignedBigInteger('planetas_id')->unsigned();
             $table->foreign('planetas_id')->references('id')->on('planetas');
         });
     }
